refactor(gallery): page.tsx is now GalleryInner and nothing else
4,974 -> 4,310. What moved was everything still sitting around the component:
379 lines of demo fixtures into _demo-items.ts, PendingGenTile into its own
file, and eleven helpers — mention resolution and rendering, settings
persistence, tag renumbering, reorder display order — onto _shared.tsx.

The three module-level `let`s behind reorder drag became a holder object, for
the same reason galleryDrag did: an imported binding cannot be reassigned, and
every one of them is written from a drag handler.

Verified as a move, not a rewrite: diffing the line multiset leaves three
differences, all comment lines the block boundaries clipped. The one carrying
meaning is restored. Cards render, the lightbox opens with its prompt, console
clean.

GalleryInner itself stays at 4,189 lines, deliberately. Three separate
measurements say it does not come apart:

  lift state into hooks   92 of 105 state/refs form ONE island in the logic
                          alone, with zero clusters of 2+ outside it
  extract the main JSX    the 1,133-line subtree needs 190 props
  extract the next one    the 357-line subtree needs 117

Every route out of it is the same coupling wearing a different hat. The real
fix is moving that state into the zustand store the workflow side already uses,
which is a rewrite of the component's state model rather than a refactor of its
layout — and not something to start days before a release.
