+++ app/gallery/_demo-items.ts
@@ -1,0 +1,371 @@
+// The sample gallery shown when NEXT_PUBLIC_DEMO_MODE is on. Pure fixture data,
+// 379 lines of it, which had no reason to sit above the page component.
+import type { GalleryItem } from "@/lib/galleryUtils";
+
+export const DEMO_GALLERY_ITEMS: GalleryItem[] = [
+  {
+    id: "demo-1",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/hf_20260421_160252_c4c544a7-07fb-4393-9449-3349e87fc597.png",
+    mediaType: "image",
+    model: "gpt-image-2",
+    quality: "medium",
+    source: "generation",
+    created_at: "2026-04-21T16:02:52.000Z",
+    prompt: `Hyper-realistic game screenshot, next-gen AAA quality,
+photorealistic third-person action RPG,
+ultra-detailed open world environment,
+
+SCENE:
+ancient Greek mountain pass, winter overcast,
+heavy grey-white sky, diffuse cold light,
+no direct sun — soft flat winter shadows,
+broken Greek stone columns half-buried in snow,
+ruined stone archway carved with meander patterns,
+distant mountain peaks disappearing into grey haze,
+snow surface worn and dirty from battle,
+muddy boot prints, dark blood spots in snow,
+scattered broken arrows half-buried,
+volumetric cold mist low to ground,
+breath vapor rising from all characters,
+loose snow drifting across ground in wind,
+depth of field blur on distant mountains,
+film grain subtle overlay,
+
+PLAYER CHARACTER:
+male hunter-archer, lean muscular build,
+photorealistic skin — red nose, chapped lips,
+cold weathered face, scar on jaw,
+subsurface scattering on skin — cold blue undertone,
+worn leather and fur armor — realistic damage,
+bear fur pauldron matted and frost-damp,
+bronze bracers tarnished green-brown,
+quiver on back — arrows worn and real,
+bowstring fully drawn, aiming upward,
+one knee pressed deep into snow,
+snow packed into boot soles and knee,
+torn cloak edge fraying in wind,
+dark hair with small braids — wet and heavy,
+finger tendons white from bowstring tension,
+micro detail: sweat, dirt, frost on eyebrows,
+dynamic combat pose, motion blur on arrow,
+
+CYCLOPS BOSS:
+single massive Cyclops — 7 meters tall,
+Argos — Stone Titan,
+photorealistic stone-grey cracked skin,
+subsurface scattering on skin — cold tone,
+wounds and cracks from prior battle,
+dried blood on knuckles and arms,
+single biological eye — natural amber,
+no magic glow — real organic texture,
+crude armor from animal hides and rusted iron,
+each hide strap and buckle individually detailed,
+rusted chains as weapon — links individually rendered,
+mid-step forward — weight shift visible,
+snow compressing and cracking under foot,
+massive exhale breath cloud — animal scale,
+expression of focused predatory rage,
+muscle and tendon anatomy under skin,
+Nanite-level skin and armor detail,
+
+COMBAT MOMENT:
+arrow frozen mid-air leaving bowstring,
+real arrow physics — no magic trail,
+fletching blur from natural spin,
+arrowhead glinting in diffuse grey light,
+Cyclops single eye tracking the arrow,
+chains mid-swing natural motion blur,
+snow spray from last footstep impact,
+player absolutely still — breath held,
+cinematic frozen tension moment,
+
+LIGHTING:
+flat diffuse grey overcast sky light,
+cold blue-grey color temperature 6500K,
+snow reflecting ambient light upward as fill,
+natural rim on player from snow bounce,
+no god rays, no bloom, no lens flare,
+no artificial glow anywhere,
+ray-traced global illumination,
+path-traced shadows soft and natural,
+ACES filmic tone mapping,
+slightly desaturated — realistic winter palette,
+volumetric cold mist catching diffuse light,
+HDR subtle — no blown highlights,
+chromatic aberration very subtle on edges,
+
+CAMERA:
+slight low angle — player foreground left,
+Cyclops dominating center-right upper frame,
+grey overcast sky filling top third,
+anamorphic 2.39:1 framing,
+shallow depth of field — f/2.8,
+background mountains soft bokeh,
+no lens flare,
+feels exactly like paused real gameplay,
+
+TECHNICAL:
+Unreal Engine 5.3, Lumen GI, Nanite geometry,
+path-traced shadows, 4K resolution,
+photorealistic texture detail 16K,
+ray-traced ambient occlusion,
+film grain subtle, chromatic aberration minimal,
+anamorphic lens bokeh on background,
+16:9 aspect ratio game screenshot,
+God of War + Ghost of Tsushima + Horizon Zero Dawn visual fidelity`,
+  },
+  {
+    id: "demo-2",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/hf_20260421_162100_a9cb557d-1397-42c1-ac16-499c09042ea5.png",
+    mediaType: "image",
+    model: "gpt-image-2",
+    quality: "medium",
+    source: "generation",
+    created_at: "2026-04-21T16:21:00.000Z",
+    prompt: `Horizontal desktop landing page hero section for a fictional snowboard eyewear and mountain gear store called "APEX & SUMMIT" (placeholder). Cinematic sports-tech advertising aesthetic with extreme close-up athletic portrait, reflection-based storytelling, bold condensed voice-command headline, premium performance-brand layout. 16:9 desktop composition.
+
+Canvas: Full-width horizontal 16:9 desktop hero. Deep snow-white to soft icy-blue gradient background with subtle motion-blur snow particles drifting across the frame.
+
+Main hero image: An extreme close-up cinematic portrait of a young professional snowboarder — a woman with her hair pulled back tightly, visible small beads of melted snow on her temples and eyebrows, sun-kissed skin with a subtle athletic glow. She wears oversized performance snowboard goggles with a bold wraparound frame and large mirrored reflective lenses tinted in a vivid amber-gold to deep magenta gradient.
+
+Key storytelling element — the reflection in her goggles: Inside the mirrored lens surface, a sharp and detailed reflection of a dynamic snowboard action scene is visible — a snowboarder mid-air executing a jump off a pristine mountain ridge against bright blue sky, powder snow spraying out behind them, mountain peaks receding into the distance.
+
+Main headline (right side, massive bold condensed sans-serif):
+"HEY APEX / CUE MY / 'FIRST DESCENT' / PLAYLIST"
+
+Sub-headline: "Smart goggles for smarter runs."
+
+Palette: icy-blue and snow-white gradient background, warm amber-to-magenta reflective lens, natural sun-kissed skin tones, crisp white typography.
+Mood: high-performance athletic, premium technology-forward, cinematic commercial, aspirational mountain sport. Awwwards / Behance site-of-the-day premium sportswear landing page, 4k.`,
+  },
+  {
+    id: "demo-3",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/hf_20260422_144420_5f74e1a3-ff9e-4f59-b8f8-799982bdd36b.png",
+    mediaType: "image",
+    model: "gpt-image-2",
+    quality: "medium",
+    source: "generation",
+    created_at: "2026-04-22T14:44:20.000Z",
+    prompt: `Contemporary commercial lifestyle photograph, luxury snow-sports campaign aesthetic, Moncler × Prada Linea Rossa × Arc'teryx × Jacquemus Ski campaign style, authentic iPhone 16 Pro aesthetic.
+
+Hero concept: a playful commercial selfie moment — a young man and young woman, both snowboarders, pause mid-slope for an intimate joyful selfie where the woman photographs her own reflection in the massive mirrored lens of the man's ski goggles.
+
+Hero subject foreground: a handsome young man in his mid-twenties, slight stubble, confident relaxed expression with a small knowing smile, wearing a ski helmet and oversized chrome-mirrored ski goggles with a reflective silver-blue lens. He wears a vivid electric-orange puffer jacket with technical stitching details.
+
+Reflected in his mirrored goggle lens: the young woman photographer reflected perfectly, holding up an iPhone 16 Pro at arm's length. She wears a hot-pink and magenta gradient ski suit with retro 90s-inspired color-blocking, white ski helmet, long dark hair spilling out in two loose braids, huge bright smile. Behind her in the reflection: snowy alpine mountain peaks, bluebird sky, other skiers, a wooden chalet lodge in the distance.
+
+Color grading: punchy commercial palette with authentic iPhone computational photography signature — electric orange jacket popping aggressively against deep cobalt-blue sky, hot pink and magenta in the reflection, chrome silver-blue mirrored goggle lens.
+
+Shot on iPhone 16 Pro, Moncler × Prada Linea Rossa × Arc'teryx ski campaign aesthetic, joyful high-energy couple, clever conceptual selfie construction, ultra-detailed, vivid saturated alpine color palette.`,
+  },
+  {
+    id: "demo-4",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/hf_20260422_211323_4cc2cdec-06a1-4d10-93b9-a63a44ae5d93.png",
+    mediaType: "image",
+    model: "gpt-image-2",
+    quality: "medium",
+    source: "generation",
+    created_at: "2026-04-22T21:13:23.000Z",
+    prompt: `Surreal cinematic portrait photograph, shot on medium-format film camera (Mamiya 7 II or Pentax 67) with 80mm lens, Kodak Portra 400 film, soft natural Mediterranean afternoon daylight, dreamy painterly color grade with cool ocean blue background and warm earthy skin tones, surrealist documentary aesthetic, Ryan McGinley meets Slim Aarons coastal-dreamlike mood.
+
+A close-up surreal portrait of a young European man in his mid-twenties, captured from behind and slightly to the side at a three-quarter back angle. He's positioned just at the water's edge of a calm Mediterranean sea or coastal harbor.
+
+He has wet sun-bleached light golden-brown hair, slightly tousled and damp from swimming. A glimpse of his face in profile is visible — a sharp masculine cheekbone with light freckles and faint stubble along his jawline. His skin is sun-tanned warm honey-bronze with subtle red flush from the summer sun. A delicate thin gold chain necklace is just visible around his neck.
+
+Draped dramatically across the top of his head and tumbling down the side of his face is a freshly caught raw octopus — a real cephalopod with a soft pinkish-coral mantle sitting balanced on top of his hair like an absurd surreal hat, and multiple long muscular tentacles cascading downward in organic spiral curves. The tentacles are a beautiful gradient of warm amber, peachy-pink, and deep umber-brown tones, the underside revealing rows of perfect circular pale-cream suction cups.
+
+The background is the calm Mediterranean ocean — a vast soft gradient of pale azure-blue at the horizon transitioning into deeper teal-blue. Painterly medium-format film grain, soft Portra color rendering with natural saturation, dreamy surrealist coastal aesthetic.`,
+  },
+  {
+    id: "demo-5",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/hf_20260422_213610_8f9804a3-5fce-48c0-94e0-5147ab7cf3fd.png",
+    mediaType: "image",
+    model: "gpt-image-2",
+    quality: "medium",
+    source: "generation",
+    created_at: "2026-04-22T21:36:10.000Z",
+    prompt: `Horizontal desktop landing page for a fictional virtual outfit try-on service called "DRESSR" (placeholder). Full Y2K / PS2-era video game character select screen aesthetic. Low-poly 3D models, glossy translucent UI panels, aqua-cyan grid backgrounds, pixelated retro typography, chrome bevels.
+
+Background: A deep aqua-cyan gradient background with visible wireframe grid lines receding into perspective. Scanline overlay and slight CRT-monitor glow throughout.
+
+Left panel: A large glossy cyan-chrome-framed display panel containing a low-poly 3D avatar of a young woman standing in a neutral T-pose, rendered in early-2000s PS2 graphics style. She has long dark braided pigtails, wearing a pink cropped tank top, low-rise denim skirt with studded belt — classic Y2K outfit.
+
+Main headline (center, massive Y2K bubble-chrome 3D typography):
+"TRY / EVERYTHING."
+
+Right panel — Character roster grid: A massive glossy 3D glass panel with a 7×5 grid of small square portrait thumbnails — each showing a different low-poly 3D avatar wearing a different outfit.
+
+CTA buttons: Two large pill-shaped glossy 3D buttons:
+- Left (magenta-pink gradient): "▶ START FITTING"
+- Right (cyan-blue gradient): "HOW IT WORKS"
+
+Palette: deep aqua-cyan (#0A4A6B to #5BC0D8), bright magenta-pink (#FF3D8E), cyan highlights (#4FE5FF), chrome silver, pixel-green LCD.
+Mood: nostalgic Y2K video game, Sega Dreamcast meets Dance Dance Revolution meets early Flash-era web. Awwwards / Dribbble-quality Y2K revival landing page, 4k.`,
+  },
+  {
+    id: "demo-6",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/hf_20260422_215112_09f9313d-4c93-40d5-8d5b-e6a502fa93c1.png",
+    mediaType: "image",
+    model: "gpt-image-2",
+    quality: "medium",
+    source: "generation",
+    created_at: "2026-04-22T21:51:12.000Z",
+    prompt: `Ultra-realistic next-gen video game screenshot, PS5 gameplay capture, exactly like real game footage, Ghost of Tsushima / Rise of the Ronin quality, NO lens flare, NO bloom, NO light rays, NO glowing effects, natural neutral lighting, overcast bright day.
+
+SCENE — JAPANESE VILLAGE STREET COMBAT:
+traditional feudal Japan village, dirt road main street, wooden buildings both sides, merchant stalls, hanging cloth banners, barrels and crates scattered, civilians fleeing in background, dust kicked up from combat movement, overcast white sky — diffuse natural light, photorealistic wood and stone textures, mud on ground, footprints visible, broken pottery on ground.
+
+PLAYER CHARACTER back view center:
+female samurai, white and silver armor, clean practical design, no fantasy elements, katana raised in combat guard stance, normal gameplay posture — not cinematic pose, hair tied back practically, armor worn but functional, NO glow on weapons, NO energy effects, just steel catching normal daylight.
+
+ENEMIES:
+Enemy 1 (directly ahead 5m): ronin swordsman, grey clothing, brown worn armor, chipped katana, aggressive forward stance mid-attack, "浪人 Ronin" white name above head, red health bar.
+Enemy 2 (left flank): spear ashigaru running in, "足軽 Guard" name tag, yellow health bar.
+Enemy 3 (right): archer nocking arrow, "弓兵 Archer" name tag, green health bar.
+
+HUD — clean minimal flat game UI with HP bars, minimap, item slots, button prompts.
+
+OVERALL: feels 100% like real PS5 gameplay video, natural camera position, overcast flat natural lighting, every detail sharp and clear, Ghost of Tsushima gameplay screenshot level, 1920x1080 raw capture.`,
+  },
+  {
+    id: "demo-7",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/hf_20260213_180040_5b054efc-50a9-4899-a36c-3505197b90f2.png",
+    mediaType: "image",
+    model: "gpt-image-2",
+    quality: "medium",
+    source: "generation",
+    created_at: "2026-02-13T18:00:40.000Z",
+    prompt: `A fashion-forward full-body editorial shot captures a young woman standing indoors against a clean, plain off-white wall, styled in an eclectic, avant-garde street couture look. The camera is positioned at eye level, straight-on, allowing the entire outfit and silhouette to be clearly visible while maintaining a natural, documentary feel.
+
+She wears a cropped gray faux-fur jacket with a shaggy texture, structured with visible straps and metal buckles that add a utilitarian edge. Underneath, she wears a crisp white button-up shirt left partially open, paired with a patterned purple tie hanging loosely down the center. A low-rise gray skirt with a soft drape sits at the hips.
+
+She wears a black headscarf fitted tightly over short dark hair, paired with futuristic wraparound sunglasses. In her hand, she carries a structured gray handbag adorned with playful charms. On her feet are pointed-toe knee-high boots in a muted lavender-pink tone, detailed with straps, studs, and layered textures.
+
+The setting is minimal and uncluttered, with a simple wooden floor and a plain wall. Lighting is soft and natural, evenly illuminating the subject without harsh shadows.
+
+The overall mood is experimental, rebellious, and fashion-centric, blending Y2K influences, cyber-streetwear, and deconstructed tailoring. Color palette is built around cool grays, muted purples, soft pinks, black accents, and subtle metallic tones.`,
+  },
+  {
+    id: "demo-8",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/hf_20260214_192051_a5f7769c-4e56-4d82-a9d5-cd2880c58439.png",
+    mediaType: "image",
+    model: "gpt-image-2",
+    quality: "medium",
+    source: "generation",
+    created_at: "2026-02-14T19:20:51.000Z",
+    prompt: `A high-angle medium shot features a young East Asian woman, possibly in her early 20s, with a short, black bob hairstyle situated against a façade featuring large window blinds. She poses in an edgy outfit consisting of a glossy, deep brown leather bomber jacket with elastic cuffs, a white corset-style top, a red striped tie, and an asymmetric, black distressed denim skirt with a frayed hemline and visible seams. A large, riveted belt cinches her waist. She accessorizes with sparkling bracelets and a rectangular black handbag with a chain strap placed beside her.
+
+The background showcases a light beige stone step and wall, with a modern glass window to her right. The lighting is natural with a soft shadow, and the color palette is muted with dark and neutral tones. Captured digitally with deep focus, giving the setting a crisp and urban aesthetic reminiscent of fashion editorial photography.`,
+  },
+  {
+    id: "demo-9",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/hf_20260218_130714_2430d3a3-b8b2-4c9e-894d-b29503ee12c0.png",
+    mediaType: "image",
+    model: "gpt-image-2",
+    quality: "medium",
+    source: "generation",
+    created_at: "2026-02-18T13:07:14.000Z",
+    prompt: `A straight-on medium shot depicts a young Asian woman, approximately in her 20s, sitting against a distressed concrete wall backdrop. She has long straight blonde hair and wears clear glasses. Her attire includes a pinstriped button-up blouse with a tie scarf detail, paired with a short black skirt. Distinctive items in her outfit are her shiny black patent knee-high boots featuring wedge heels.
+
+The concrete wall behind her is aged and weathered, contributing to an industrial yet grunge atmosphere. The lighting is diffuse, seemingly natural ambient light filtering in, casting soft shadows and highlighting her smooth skin. The color palette consists of neutral tones, including her clothing and the beige of the concrete, with slight accents from her hair and glasses. The image emphasizes clarity in details yet maintains a slightly muted urban and edgy mood.`,
+  },
+  {
+    id: "demo-10",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/hf_20260218_140039_e4358650-a940-43d9-90e6-3d07a9b9c7bd.png",
+    mediaType: "image",
+    model: "gpt-image-2",
+    quality: "medium",
+    source: "generation",
+    created_at: "2026-02-18T14:00:39.000Z",
+    prompt: `A medium shot captures the midsection of a young woman with a fair skin tone, showcasing edgy, alternative fashion. Her attire features an oversized, distressed black knit crop top adorned with metallic pyramid studs forming an abstract pattern. She pairs it with high-waisted, loose-fitting gray denim pants adorned with a black belt matching silver studs, and stacks of metallic and black cuff bracelets on her wrists.
+
+The backdrop consists of a sheer white curtain, creating a minimalist, neutral setting. The lighting is soft and ambient, producing gentle shadows and highlights. The color palette is focused on blacks, grays, and metallics, conveying a modern and edgy atmosphere.`,
+  },
+  {
+    id: "demo-11",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/hf_20260218_142536_7eadd267-55c2-43b3-acb5-a748230ca3ad.png",
+    mediaType: "image",
+    model: "gpt-image-2",
+    quality: "medium",
+    source: "generation",
+    created_at: "2026-02-18T14:25:36.000Z",
+    prompt: `An overhead close-up shot captures the hands of an individual with light skin tone, showcasing long stiletto-shaped acrylic nails. The nails are a gradient of dusty pink, transitioning to beige tips, emphasizing an elongated and dramatic shape. The hands have visible tattoos, including intricate linework and geometric designs. Silver rings adorn fingers on both hands, adding a metallic contrast to the natural hues. The person's sleeves have a lace edge detail, contributing a delicate texture against the boldness of the nails.
+
+Lighting is soft and diffused, reducing shadows and lending the image a flat, almost clinical aesthetic. The color palette is predominantly neutral with beige, pink, and silver tones. The composition centers the hands against a plain, light gray backdrop, enhancing focus on the nail art and tattoos. The image is sharp with fine details on the skin and fabric textures, creating a bold and edgy atmosphere.`,
+  },
+];
+
+export const DEMO_VIDEO_ITEMS: GalleryItem[] = [
+  {
+    id: "demo-v1",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/1.mp4",
+    mediaType: "video",
+    model: "Seedance 2.0",
+    source: "generation",
+    created_at: "2026-04-22T10:00:00.000Z",
+    prompt: `A young skateboarder in a black t-shirt and baggy jeans performs tricks at a brutalist concrete skatepark on a sunny afternoon. The camera alternates between a fisheye lens ground-level POV following his board and wider angles capturing kickflips and grinds. Harsh midday shadows, raw street-skating documentary style, warm golden tones.`,
+  },
+  {
+    id: "demo-v2",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/2.mp4",
+    mediaType: "video",
+    model: "Seedance 2.0",
+    source: "generation",
+    created_at: "2026-04-22T10:05:00.000Z",
+    prompt: `Extreme macro close-up of ornate stone chess pieces on a wooden chessboard inside a stylish mid-century modern living room. A hand reaches in to move a piece. The camera slowly descends to board level, revealing the intricate carved detail of a king and queen piece towering like monumental sculptures. Warm interior lighting, shallow depth of field, elegant product-cinematography aesthetic.`,
+  },
+  {
+    id: "demo-v3",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/3.mp4",
+    mediaType: "video",
+    model: "Seedance 2.0",
+    source: "generation",
+    created_at: "2026-04-22T10:10:00.000Z",
+    prompt: `A man in a dark hoodie stands in the center of a vast industrial facility filled with metal pipes and overhead fluorescent strip lights. His body begins to morph and assemble into a sleek dark robotic exosuit. The camera pulls back to reveal fallen mechanical wreckage on the floor around him. Dark sci-fi atmosphere, cold blue-grey tones, dramatic low-angle transformation sequence.`,
+  },
+  {
+    id: "demo-v4",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/4.mp4",
+    mediaType: "video",
+    model: "Seedance 2.0",
+    source: "generation",
+    created_at: "2026-04-22T10:15:00.000Z",
+    prompt: `Underwater close-up inside a planted freshwater aquarium. A piece of raw meat drifts down through the water. Two large piranhas with dark bodies and orange-red bellies surge forward and tear into the meat aggressively, pulling it apart. Driftwood and lush green aquatic plants fill the background. Crystal-clear water, nature-documentary macro photography, dramatic feeding behavior.`,
+  },
+  {
+    id: "demo-v5",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/5.mp4",
+    mediaType: "video",
+    model: "Seedance 2.0",
+    source: "generation",
+    created_at: "2026-04-22T10:20:00.000Z",
+    prompt: `Close-up portrait of a young woman with wavy honey-blonde hair walking on a city street at night. She holds a coffee cup in one hand and a smartphone in the other, looking at the screen with a pouty expression. She wears a navy blue and white striped athletic jacket. Out-of-focus city lights and car headlights create warm red and orange bokeh in the background. The camera slowly orbits around her face. Nighttime urban setting, shallow depth of field, warm skin tones lit by the phone screen glow, candid street photography style, cinematic close-up.`,
+  },
+  {
+    id: "demo-v6",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/6.mp4",
+    mediaType: "video",
+    model: "Seedance 2.0",
+    source: "generation",
+    created_at: "2026-04-22T10:25:00.000Z",
+    prompt: `Cinematic tracking shot following a young woman with long pink hair walking through a high school hallway lined with lockers. She wears a white cropped zip-up hoodie and carries a black backpack over one shoulder, looking down pensively. Other students move in the background. The camera slowly orbits from a front-facing medium shot to behind her as she continues walking. Warm tungsten overhead lighting, shallow depth of field, 2000s teen drama aesthetic, shot on anamorphic lens with subtle lens flares.`,
+  },
+  {
+    id: "demo-v7",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/7.mp4",
+    mediaType: "video",
+    model: "Seedance 2.0",
+    source: "generation",
+    created_at: "2026-04-22T10:30:00.000Z",
+    prompt: `Inside a sandy desert terrarium with dried branches and a small cactus, a large hairy tarantula and an orange scorpion converge on a piece of raw pink meat placed on the substrate. They grapple over the food in a tense confrontation. Warm overhead lighting, extreme macro close-up, nature-documentary tension, detailed textures of chitin and sand.`,
+  },
+  {
+    id: "demo-v8",
+    url: "https://pub-73a59b956f1c4a7db2934522c13d8027.r2.dev/demo-mode/8.mp4",
+    mediaType: "video",
+    model: "Seedance 2.0",
+    source: "generation",
+    created_at: "2026-04-22T10:35:00.000Z",
+    prompt: `A young woman with bright orange-red hair and chunky over-ear headphones sits at an airport terminal gate by a floor-to-ceiling window, eating sushi from a small plate. She wears a sporty racing-style jacket. Travelers pull luggage past her in the background. Natural daylight streams through the glass, casual candid moment, shallow depth of field, travel lifestyle aesthetic.`,
+  },
+];
